fix: fix burst-mode errors + switch app from kapt to KSP
burst-mode (7 errors fixed):
- BurstModeImpl: fix Long->Int with .toInt() on estimateTokens
- CheckpointManager: add name+input params to BurstTask + import BurstInput
- TaskContext: remove inline from recursive getOrNull, fix get<Long> ambiguity
  by using direct cast (variables[key] as? Long)

app module:
- Switch from kapt to KSP (KSP supports Kotlin 2.0+, kapt doesn't)
- Replace kapt(libs.google.hilt.compiler) -> ksp(...)
- Replace kapt(libs.androidx.room.compiler) -> ksp(...)
- Disable ObjectBox processor (KSP not supported by ObjectBox)
- kapt block -> ksp block for room.schemaLocation arg

Expected: 9/9 APKs build successfully

+++ core/burst-mode/src/main/java/com/apex/agent/burstmode/api/BurstModeImpl.kt
@@ -542,7 +542,7 @@ internal class BurstModeImpl(
             output = "BurstMode executed task: ${task.description.take(200)}",
             metrics = com.apex.agent.plugins.burst.base.SkillMetrics(
                 executionTimeMs = 0,
-                tokensProcessed = estimateTokens(task.description)
+                tokensProcessed = estimateTokens(task.description).toInt()
             )
         )
     }
